backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
sse-starlette==1.6.5
websockets==12.0

# LLM Provider SDKs
openai==1.3.5
anthropic==0.7.8

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Configuration and Environment
pydantic>=2.9.0
pydantic-settings>=2.6.0
pyyaml==6.0.1
python-dotenv==1.0.0

# MCP Protocol
# mcp==0.9.0  # Commented out - will be added when MCP integration is implemented

# Utilities
typing-extensions==4.8.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Logging
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

